requirements.txt
fastapi
uvicorn
python-dotenv
twilio
aiofiles
loguru
pipecat-ai[google,silero]
python-multipart
pipecat-ai[deepgram]

